requirements.txt
pandas
selenium
groq
langchain
langchain-community
langchain_groq
faiss-cpu
sentence-transformers
langchain_huggingface
flask
flask[async]